Python in ArcGIS Pro
use Python in ArcGIS Pro through multiple access locations to process infrastructure data. A map of industrial buildings and highways for the county, as well as school and sewer-line data for the state. 1) Clip the schools and sewer-line data using the Providence county border; 2) Update highway field values using the field calculator.