Update Use the Python window for raster analysis.py
Python window to develop a script to model a peregrine falcon habitat. Identify areas to set aside within the park for the peregrine falcon. These birds are considered a threatened species so consider potential habitat sites. The script is used to model the areas within the park that meet the habitat criteria.